requirements.txt
cryptography
pillow
numpy
tk